requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
colorama==0.4.3
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
entrypoints==0.3
gunicorn==20.0.4
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
Jinja2==2.10.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.1.0
jupyter-core==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==5.7.8
numpy==1.14.2
olefile==0.46
pandocfilters==1.4.2
parso==0.5.2
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.4
psycopg2-binary==2.8.4
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.0
qtconsole==4.6.0
six==1.13.0
sqlparse==0.3.0
testpath==0.4.4
tornado==5.1.1
traitlets==4.3.3
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==5.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==0.6.0
django-taggit==1.2.0
django-ckeditor==5.9.0
cloudinary==1.20.0
django-cloudinary-storage==0.2.3